+++ Tests/ConvertTo-UnixTime.Tests.ps1
@@ -1,4 +1,4 @@
-Describe $($PSCommandPath -Replace ".Tests.ps1") {
+Describe $($PSCommandPath -Replace '.Tests.ps1') {
 
 	BeforeAll {
 		#Get Current Directory
@@ -20,8 +20,8 @@ Describe $($PSCommandPath -Replace ".Tests.ps1") {
 		}
 
 		$Script:RequestBody = $null
-		$Script:BaseURI = "https://SomeURL/SomeApp"
-		$Script:ExternalVersion = "0.0"
+		$Script:BaseURI = 'https://SomeURL/SomeApp'
+		$Script:ExternalVersion = '0.0'
 		$Script:WebSession = New-Object Microsoft.PowerShell.Commands.WebRequestSession
 
 	}
@@ -35,11 +35,11 @@ Describe $($PSCommandPath -Replace ".Tests.ps1") {
 
 	InModuleScope $(Split-Path (Split-Path (Split-Path -Parent $PSCommandPath) -Parent) -Leaf ) {
 
-		Context "Mandatory Parameters" {
+		Context 'Mandatory Parameters' {
 
 			$Parameters = @{Parameter = 'Date' }
 
-			It "specifies parameter <Parameter> as mandatory" -TestCases $Parameters {
+			It 'specifies parameter <Parameter> as mandatory' -TestCases $Parameters {
 
 				param($Parameter)
 
@@ -49,27 +49,43 @@ Describe $($PSCommandPath -Replace ".Tests.ps1") {
 
 		}
 
-		Context "General" {
+		Context 'General' {
 
-			It "Converts date from pipeline" {
+			It 'Converts date from pipeline' {
 				$(Get-Date -Year 2020 -Month 01 -Day 01 -Hour 0 -Minute 0 -Second 0 -Millisecond 0) | ConvertTo-UnixTime | Should -Be 1577836800
 			}
 
-			It "Converts date" {
+			It 'Converts date' {
 				$date = $(Get-Date -Year 2020 -Month 01 -Day 01 -Hour 0 -Minute 0 -Second 0 -Millisecond 0)
 				ConvertTo-UnixTime -Date $date | Should -Be 1577836800
 			}
 
-			It "converts date to expected unixtime" {
+			It 'converts date to expected unixtime' {
 
 				ConvertTo-UnixTime -Date $(Get-Date 1/1/2020) | Should -Be 1577836800
 
 			}
 
-			It "converts date to expected unixtime in milliseconds" {
+			It 'converts date to expected unixtime in milliseconds' {
 
 				ConvertTo-UnixTime -Date $(Get-Date 1/1/2020) -Milliseconds | Should -Be 1577836800000
 
+			}
+
+			It 'converts date to expected unixtime in milliseconds, even when locale is not en-US' {
+				$currentCulture = [System.Threading.Thread]::CurrentThread.CurrentCulture
+				[System.Threading.Thread]::CurrentThread.CurrentCulture = 'nl-NL'
+				ConvertTo-UnixTime -Date $(Get-Date 1/1/2020) -Milliseconds | Should -Be 1577836800000
+				[System.Threading.Thread]::CurrentThread.CurrentCulture = $currentCulture
+			}
+
+			It 'converts date to expected unixtime in milliseconds, even when locale is "SomeLocale"' {
+				$currentCulture = [System.Threading.Thread]::CurrentThread.CurrentCulture
+				[System.Threading.Thread]::CurrentThread.CurrentCulture = $(Get-Culture -ListAvailable |
+						Select-Object -Index (Get-Random -Maximum $((Get-Culture -ListAvailable).Count)) |
+						Select-Object -ExpandProperty Name)
+				ConvertTo-UnixTime -Date $(Get-Date 1/1/2020) -Milliseconds | Should -Be 1577836800000
+				[System.Threading.Thread]::CurrentThread.CurrentCulture = $currentCulture
 			}
 
 		}