+++ psPAS/Functions/Accounts/Get-PASAccountPassword.ps1
@@ -1,77 +1,83 @@
 # .ExternalHelp psPAS-help.xml
 function Get-PASAccountPassword {
-	[CmdletBinding(DefaultParameterSetName = "Gen2")]
+	[CmdletBinding(DefaultParameterSetName = 'Gen2')]
 	param(
 		[parameter(
 			Mandatory = $true,
 			ValueFromPipelinebyPropertyName = $true,
-			ParameterSetName = "Gen1"
+			ParameterSetName = 'Gen1'
 		)]
 		[parameter(
 			Mandatory = $true,
 			ValueFromPipelinebyPropertyName = $true,
-			ParameterSetName = "Gen2"
+			ParameterSetName = 'Gen2'
 		)]
-		[Alias("id")]
+		[Alias('id')]
 		[string]$AccountID,
 
 		[parameter(
 			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $false,
-			ParameterSetName = "Gen1"
+			ParameterSetName = 'Gen1'
 		)]
-		[Alias("UseClassicAPI")]
+		[Alias('UseClassicAPI')]
 		[switch]$UseGen1API,
 
 		[parameter(
 			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $false,
-			ParameterSetName = "Gen2"
+			ParameterSetName = 'Gen2'
 		)]
 		[string]$Reason,
 
 		[parameter(
 			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $false,
-			ParameterSetName = "Gen2"
+			ParameterSetName = 'Gen2'
 		)]
 		[string]$TicketingSystem,
 
 		[parameter(
 			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $false,
-			ParameterSetName = "Gen2"
+			ParameterSetName = 'Gen2'
 		)]
 		[string]$TicketId,
 
 		[parameter(
 			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $false,
-			ParameterSetName = "Gen2"
+			ParameterSetName = 'Gen2'
 		)]
 		[int]$Version,
 
 		[parameter(
 			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $false,
-			ParameterSetName = "Gen2"
+			ParameterSetName = 'Gen2'
 		)]
-		[ValidateSet("show", "copy", "connect")]
+		[ValidateSet('show', 'copy', 'connect')]
 		[string]$ActionType,
 
 		[parameter(
 			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $false,
-			ParameterSetName = "Gen2"
+			ParameterSetName = 'Gen2'
 		)]
 		[boolean]$isUse,
 
 		[parameter(
 			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $false,
-			ParameterSetName = "Gen2"
+			ParameterSetName = 'Gen2'
 		)]
-		[switch]$Machine
+		[string]$Machine,
+
+		[parameter(
+			Mandatory = $false,
+			ValueFromPipelinebyPropertyName = $true
+		)]
+		[String]$UserName
 	)
 
 	BEGIN {
@@ -83,19 +89,19 @@ function Get-PASAccountPassword {
 		#Build Request
 		switch ($PSCmdlet.ParameterSetName) {
 
-			"Gen2" {
+			'Gen2' {
 
 				Assert-VersionRequirement -RequiredVersion 10.1
 
 				#For Version 10.1+
 				$Request = @{
 
-					"URI"    = "$Script:BaseURI/api/Accounts/$($AccountID | Get-EscapedString)/Password/Retrieve"
+					'URI'    = "$Script:BaseURI/api/Accounts/$($AccountID | Get-EscapedString)/Password/Retrieve"
 
-					"Method" = "POST"
+					'Method' = 'POST'
 
 					#Get all parameters that will be sent in the request
-					"Body"   = $PSBoundParameters | Get-PASParameter -ParametersToRemove AccountID | ConvertTo-Json
+					'Body'   = $PSBoundParameters | Get-PASParameter -ParametersToRemove AccountID, UserName | ConvertTo-Json
 
 				}
 
@@ -103,14 +109,14 @@ function Get-PASAccountPassword {
 
 			}
 
-			"Gen1" {
+			'Gen1' {
 
 				#For Version 9.7+
 				$Request = @{
 
-					"URI"    = "$Script:BaseURI/WebServices/PIMServices.svc/Accounts/$($AccountID | Get-EscapedString)/Credentials"
+					'URI'    = "$Script:BaseURI/WebServices/PIMServices.svc/Accounts/$($AccountID | Get-EscapedString)/Credentials"
 
-					"Method" = "GET"
+					'Method' = 'GET'
 
 				}
 
@@ -121,7 +127,7 @@ function Get-PASAccountPassword {
 		}
 
 		#Add default Request parameters
-		$Request.Add("WebSession", $Script:WebSession)
+		$Request.Add('WebSession', $Script:WebSession)
 
 		#splat request to web service
 		$result = Invoke-PASRestMethod @Request
@@ -130,7 +136,7 @@ function Get-PASAccountPassword {
 
 			switch ($PSCmdlet.ParameterSetName) {
 
-				"Gen1" {
+				'Gen1' {
 
 					$result = [System.Text.Encoding]::ASCII.GetString([PSCustomObject]$result.Content)
 
@@ -138,10 +144,18 @@ function Get-PASAccountPassword {
 
 				}
 
-				"Gen2" {
+				'Gen2' {
 
 					#Unescape returned string and remove enclosing quotes.
 					$result = $([System.Text.RegularExpressions.Regex]::Unescape($result) -replace '^"|"$', '')
+
+					#Get UserName if parameter value not provided.
+					if ($PSBoundParameters.Keys -notcontains 'UserName') {
+
+						$UserName = Get-PASAccount -id $AccountID -ErrorAction SilentlyContinue |
+							Select-Object -ExpandProperty UserName
+
+					}
 
 					break
 
@@ -149,7 +163,10 @@ function Get-PASAccountPassword {
 
 			}
 
-			[PSCustomObject] @{"Password" = $result } | Add-ObjectDetail -typename psPAS.CyberArk.Vault.Credential
+			[PSCustomObject]@{
+				'Password' = $result
+				'UserName' = $UserName
+			} | Add-ObjectDetail -typename psPAS.CyberArk.Vault.Credential
 
 		}